+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Added test for bin heights of histograms (@ryla5068, #124)
 
 ## [0.1.0]
-* Added test for bin heights of histograms (@ryla5068, #124)
 * Added support for overlapping histograms in histogram tests (@ryla5068, #123)
 * Created a vignette covering base plot tester functionality (@ryla5068, #122)
 * fix pip version to ensure pyproj installs in black tox envt (@lwasser, #144)